refactor(workflow): splitting workflow into multiple files
