+++ pkg/workflow/node_input.go
@@ -1,0 +1,4 @@
+package workflow
+
+// NodeInput node input type
+type NodeInput map[string]any
